[routes] Fix repoProxy to potentially always prefix with artifact path
This is useful for the reference documentation which has a root `docs`
directory, which we always want to dive into. As a result, the rewriting
of `../reference` to `../reference/docs/index.html` can be simplified by
removing the explicit `docs/`.

Relative urls for reference guide assets like highlight scripts should
now work.