Add /maintenance page driven by maintenanceXxx env vars
The environment variables change the message to indicate there is an
expected maintenance window:
 - `maintenanceDate` is the YYYY/MM/DD date of the scheduled maintenance
 - `maintenanceEnd` is the expected end time of said maintenance

Both are Strings interpolated in the message, so there's no format
enforced.

The page is kept minimal (no external css or scripts) as it is
intended to be cached for the CDN to serve in case of errors, and
the CDN puts a size limit of such a page.