[build] Add Spotless plugin to enforce license headers (#68)
This commit introduces Spotless plugin and configures it to allow
checking and enforcing copyright notice headers via a template, for
java files only.

The license check is added to the ci workflow, ensuring only files
touched by a PR will get checked (in order to allow incremental updates
to the end of the year range).

The build can be parameterized to deactivate the ratchet feature ("only
changed files" vs "all files") or choose a difference base branch for
comparison.

See reactor/reactor#701.