Improve discoverability of security policy, add security-txt
This commit changes the URL for /security-policy to one on VMware site
that does indeed describe the full security policy.

It adds a link to /security-policy in the site's footer.

It also introduces a /.well-known/security.txt file that conforms to the
draft security-txt spec, see https://securitytxt.org/.

Fixes #69.