Use Docker Hub read-only token (#137)
The cf-cli-action GH action used to deploy the site to cloud foundry is using Docker, but since the self-runner is inside a private network, it is sharing the gateway IP address. When too many unauthenticated docker pull requests come from the same IP, it will get temporarily blocked by Docker, and the DEPLOY workflow fails with a "toomanyrequests: You have reached your pull rate limit" error.

To work around, a Docker Hub read-only token has been created in the form of a GitHub Secret that can be used when logging to Docker Hub.