Fix buildpacks in CF manifest (#139)
When running the DEPLOY workflow in the production site, it does not work and we have the following error in the workflow logs:

 [Buildpack]                      ERROR Finalize failed with exception #<RuntimeError: Unable to find cached file for https://java-buildpack.cloudfoundry.org/jvmkill/bionic/x86_[64]

it seems to be related to the Java Buildpack not being able to find a cached file.
in Bamboo, the buildpack used is https://github.com/cloudfoundry/java-buildpack.git, but we are currently using java_buildpack_offline. Let's configure the buildpacks to https://github.com/cloudfoundry/java-buildpack.git and use the same bildpacks that is used by Bamboo.