Create Azure App And Configure Jvm Size (#153)
This PR improves the deploy-to-azure workflow for the projectreactor site Azure spring app. Previously, it was assumed that the azure spring app was manually created before using this workflow. Now, if the app doesn't exist, this PR creates it and ensures it has at least 2 GB of memory to run.

Additionally, the spring app is now deployed with Java 11, like our old CF servers. The JVM memory is configured to 1 GB as before. However, due to memory requirements, the spring app memory is set to 2 GB. Setting the JVM to 1 GB actually needs more memory. If we set the application max memory size to 1 GB, it leads to a deployment error.